components refactor
