+++ components/index.js
@@ -1,0 +1,3 @@
+export { default as FooterComponent} from './footer';
+export { default as NavbarComponent} from './navbar';
+export { default as SpinnerComponent} from './spinner';
